cleanup & add chunking pattern to user anchor

+++ zomes/chunked_social_context/src/inputs.rs
@@ -10,14 +10,14 @@ pub struct DnaAddress {
 
 #[derive(Clone, Serialize, Deserialize, SerializedBytes)]
 pub struct PaginationArguments {
-    pub from_chunk: usize,
-    pub to_chunk: usize,
+    pub from_chunk: u32,
+    pub to_chunk: u32,
 }
 
 #[derive(Clone, Serialize, Deserialize, SerializedBytes, Debug)]
 pub struct ReadCommunicationArguments {
     pub by_dna: Option<DnaHash>,
     pub by_agent: Option<Identity>,
-    pub from_chunk: usize,
-    pub to_chunk: usize,
+    pub from_chunk: u32,
+    pub to_chunk: u32,
 }
